improvement: Typing Indicator relay-side expiration Previously, there was a "buildup" of old Typing Indicators due to there being no way for relays to know that it was made to expire.
Since adding a public expiry CAN reveal metadata and typing patterns - I've decided to setup a 'Trusted Relay' that handles events like Typing Indicators, with the ability to expire them, without knowing the contents of the giftwrap.

This seems to be the most efficient half-way point that I can see, for maintaining both Privacy and Efficiency.