improvement: OS detection + Android-specific messaging fixes Vector's backend now exposes a universal "feature list" to the frontend, which allows the UI to adapt components for particular OS'es - Android needs a different method to detect keyboard input, which this system now gracefully handles.
Similarly, transcriptions cannot be done on Android, and this feature system now handles the disabling of the Transcription UI accordingly.

Lastly; the Android manifest was updated to prevent the keyboard from overlapping with the app UI - it instead resizes the app according to the space left by your keyboard, which is a significant UX improvement!